Finally applied the infamous long awaited DIV patch. ;)
Notification system enhancements to provide information about
operational type of current and previous Group node and
multi-field changes. SoDataSensor enhancements to allow querying
of aforementioned notification system changes.

Thanks to the Studierstube team for the original patch that came
out of the Distributed Inventor and to Mathis Csisinko and
Hannes Kaufmann for the reworked cleaned up version of the patch.